Group the sources into core/ and theories/, add camadatypes.h
Splits src/ by who needs a file rather than by what it is.

src/ keeps the public vocabulary: camada.h and the types its signatures
speak in. SMTExprRef appears 253 times in camada.h, so camadaexpr.h and
camadasort.h belong beside it, not one level down.

core/ holds the implementation -- camadaimpl.{h,cpp}, camada.cpp,
camadaexpr.cpp, camadasort.cpp. camadaimpl.h stays installed, as
<camada/core/camadaimpl.h>: nothing in the consumer or the tests includes
it, because it is the base class an out-of-tree backend inherits, and the
path now says so.

theories/ holds the camada-owned encodings: FP-over-BV, fixed point,
Ackermann arrays, tuple lowering. That grouping matches how the docs
already talk about them -- inside a camada-owned theory there is no
backend.

camadaenums.h becomes camadatypes.h and absorbs SMTError, SMTResult and
ArrayModel. Keeping SMTErrorCode in one file and the SMTError that wraps
it in another was an artificial line. camada.h is now 1149 lines, down
from 1433.

camadacommon.h stays as it is. It is 78 lines of compiler-attribute
macros, fatalError, and one bit-string helper, included by 17 files that
mostly want fatalErrorIf alone -- small and stable enough that a split
would cost more churn than it returns.

Verified from a clean configure: 233 tests pass, and every installed
header compiles standalone from an empty consumer include root.
