+++ src/solvers/bitwuzlasolver.h
@@ -31,7 +31,7 @@ extern "C" {
 }
 
 #include "../camadaexpr.h"
-#include "../camadaimpl.h"
+#include "../core/camadaimpl.h"
 #include "../camadasort.h"
 
 namespace camada {
