+++ src/solvers/mathsatsolver.h
@@ -31,7 +31,7 @@
 #include <variant>
 
 #include "../camadaexpr.h"
-#include "../camadaimpl.h"
+#include "../core/camadaimpl.h"
 #include "../camadasort.h"
 
 namespace camada {
