+++ src/solvers/smtlibsolver.h
@@ -31,7 +31,7 @@
 #include <vector>
 
 #include "../camadaexpr.h"
-#include "../camadaimpl.h"
+#include "../core/camadaimpl.h"
 #include "../camadasort.h"
 
 namespace camada {
