+++ src/solvers/stpsolver.h
@@ -30,7 +30,7 @@ namespace STP {
 }
 
 #include "../camadaexpr.h"
-#include "../camadaimpl.h"
+#include "../core/camadaimpl.h"
 #include "../camadasort.h"
 
 namespace camada {
