+++ src/solvers/yicessolver.h
@@ -30,7 +30,7 @@
 #include <yices.h>
 
 #include "../camadaexpr.h"
-#include "../camadaimpl.h"
+#include "../core/camadaimpl.h"
 #include "../camadasort.h"
 
 namespace camada {
