+++ src/theories/camadatuple.cpp
@@ -21,8 +21,8 @@
 
 #include "camadatuple.h"
 
-#include "camadacommon.h"
-#include "camadaimpl.h"
+#include "../camadacommon.h"
+#include "../core/camadaimpl.h"
 
 #include <algorithm>
 #include <utility>
